added E06 and C02 signs selection

+++ current_traffic_signs/select_nearest_panos.py
@@ -20,7 +20,7 @@ def pg_connection():
 def list_to_pgarray(alist):
     return "','".join(alist)
 
-def select_nearest_pano_by_sign(traffic_sign_code, fixation):
+def select_nearest_pano_by_sign(traffic_sign_code, fixation, outputformat):
     table_output_name = "nearest_panos_{}_{}".format(
         traffic_sign_code.lower(),
         fixation.split(',')[0].replace(' ', '_').lower())
@@ -29,36 +29,58 @@ def select_nearest_pano_by_sign(traffic_sign_code, fixation):
 
     print(fixations)
     cursor = pg_connection()
+    if outputformat == 'json':
+        cursor.execute(sql.SQL("""
+            DROP TABLE IF EXISTS {0};
+            SELECT DISTINCT ON (c.id)
+                c.id,
+                c.stadsdeel,
+                c.mdlnr,
+                c.mslink,
+                d.id as pano_id,
+                REPLACE(d.url,'8000','2000') as url,
+                ST_Transform(c.geom,4326) as geom
+            INTO {0}
+            FROM
+                public.panorama_recent_2017 as d,
+                (SELECT DISTINCT
+                    id, stadsdeel, bord_code, mslink, mapid, mdlnr, reflectie, bevestiging,
+                    afmeting, status, onderbordmdlnr, onderbordtekst, datum_plaats,
+                    datum_contr, datum_vervangen, wegvaknummer,
+                    opmerking, datum_vernieuw,
+                    ST_X(ST_TRANSFORM(geom,4326)) as lat,
+                    ST_Y(ST_TRANSFORM(geom,4326)) as lon,
+                    geom
+                 FROM
+                    public.current_traffic_signs
+                 WHERE "bord_code" = {1}) as c
+            WHERE ST_CONTAINS(ST_BUFFER(c.geom, 6), d.wkb_geometry)""").format(
+            sql.Identifier(table_output_name),
+            sql.Literal(traffic_sign_code)),)
+    if outputformat == 'csv':
+        cursor.execute(sql.SQL("""
+            DROP TABLE IF EXISTS {0};
+            SELECT DISTINCT ON (c.id)
+                d.*
+            INTO {0}
+            FROM
+                public.panorama_recent_2017 as d,
+                (SELECT DISTINCT
+                    id, stadsdeel, bord_code, mslink, mapid, mdlnr, reflectie, bevestiging,
+                    afmeting, status, onderbordmdlnr, onderbordtekst, datum_plaats,
+                    datum_contr, datum_vervangen, wegvaknummer,
+                    opmerking, datum_vernieuw,
+                    ST_X(ST_TRANSFORM(geom,4326)) as lat,
+                    ST_Y(ST_TRANSFORM(geom,4326)) as lon,
+                    geom
+                 FROM
+                    public.current_traffic_signs
+                 WHERE "bord_code" = {1}) as c
+            WHERE ST_CONTAINS(ST_BUFFER(c.geom, 6), d.wkb_geometry)""").format(
+            sql.Identifier(table_output_name),
+            sql.Literal(traffic_sign_code)),)
 
-    cursor.execute(sql.SQL("""
-        DROP TABLE IF EXISTS {0};
-        SELECT DISTINCT ON (c.id)
-            c.id,
-            c.stadsdeel,
-            c.mdlnr,
-            c.mslink,
-            d.id as pano_id,
-            REPLACE(d.url,'8000','2000') as url,
-            ST_Transform(c.geom,4326) as geom
-        INTO {0}
-        FROM
-            public.panorama_recent_2017 as d,
-            (SELECT DISTINCT
-                id, stadsdeel, bord_code, mslink, mapid, mdlnr, reflectie, bevestiging,
-                afmeting, status, onderbordmdlnr, onderbordtekst, datum_plaats,
-                datum_contr, datum_vervangen, wegvaknummer,
-                opmerking, datum_vernieuw,
-                ST_X(ST_TRANSFORM(geom,4326)) as lat,
-                ST_Y(ST_TRANSFORM(geom,4326)) as lon,
-                geom
-             FROM
-                public.current_traffic_signs
-             WHERE "bord_code" = {1} AND "bevestiging" = {2} OR "bord_code" = {1} AND "bevestiging" = {3}) as c
-        WHERE ST_CONTAINS(ST_BUFFER(c.geom, 6), d.wkb_geometry)""").format(
-        sql.Identifier(table_output_name),
-        sql.Literal(traffic_sign_code),
-        sql.Literal(fixations[0]),
-        sql.Literal(fixations[1])),)
+
 
     #pg_str = psycopg_connection_string
     #execute_sql(pg_str, sql_select_panos)
@@ -72,12 +94,13 @@ def parser():
     """
     description = """
     Select the nearest pano images of a specific traffic sign, save them as a table and output them as a csv file.
+    CSV outputs only the pano data, JSON outputs the registered traffic sign and the url to the pano view.
 
     Example command line:
-        ``python select_nearest_panos.py config.ini dev D02 Gele\ koker,Scharnierbeugel ../output``
+        ``python select_nearest_panos.py config.ini dev D02 Gele\ koker,Scharnierbeugel ../output csv``
 
     Result:
-        Table: nearest_panos_d02ro_bb22_gele_koker
+        Table: nearest_panos_d02ro_bb22_gele_koker_yyyy-mm-dd.csv
     """
 
     parser = argparse.ArgumentParser(
@@ -97,17 +120,26 @@ def parser():
     parser.add_argument('output_folder',
                         type=str,
                         help='Define output folder for example ../output')
+    parser.add_argument('format',
+                        type=str,
+                        help='Define output format for example json or csv')
     return parser
 
 
 def main():
     args = parser().parse_args()
-    table = select_nearest_pano_by_sign(args.traffic_sign_code, args.fixation)
-    write_table_to_geojson(
-        args.full_config_path,
-        args.db_config, table,
-        args.output_folder)
-    logger.info("Written {}/{}.json".format(args.output_folder, table))
+    table = select_nearest_pano_by_sign(args.traffic_sign_code, args.fixation, args.format)
+    if args.format == 'json':
+        write_table_to_geojson(
+            args.full_config_path,
+            args.db_config, table,
+            args.output_folder)
+    if args.format == 'csv':
+        export_table_to_csv(
+            args.full_config_path,
+            args.db_config, table,
+            args.output_folder)
+    logger.info("Written {}/{}.{}".format(args.output_folder, table, args.format))
 
 
 if __name__ == "__main__":
